META data initially scraped from fusesoc
We go through a few conversion;  .ini files are parsed into json
using python libs.  the initial json is loaded into ocaml and
then converted to an easier to grok json format which is much
less dynamically typed.  I think nearly everything survives the
conversions now but that will need more checking.

The conversions create a bunch of json files which can be easily
processed with yojson/ocaml.  The intention is to write a some
library functions and an application to query the generated metadata
along the lines of ocamlfind.