add install command
we now have a seperate step where the metadata gets installed in
a 'known' location.  The path to the actual package is added at this
point.

note; metadata is provided by a project and managed by hardcamlfind.
The project itself (ie rtl etc) is NOT going to be managed by
hardcamlfind, though it's location will be copied into the metadata
at install time.